requirements.txt
django
djangorestframework
django-crispy-forms
django-cors-headers